+++ package.json
@@ -19,14 +19,9 @@
     "discord-command-parser": "^1.5.3",
     "discord.js": "discordjs/discord.js",
     "dotenv": "^8.2.0",
-    "ejs": "^3.1.6",
     "esm": "^3.2.25",
-    "express": "^4.17.1",
     "fs-extra": "^9.1.0",
-    "json5": "^2.2.0",
-    "knex": "^0.95.2",
-    "node-cron": "^3.0.0",
-    "sqlite3": "^5.0.2"
+    "node-cron": "^3.0.0"
   },
   "devDependencies": {
     "nodemon": "^2.0.7"
